+++ libspeex/nb_celp.h
@@ -185,7 +185,7 @@ void *nb_encoder_init(SpeexMode *m);
 void nb_encoder_destroy(void *state);
 
 /** Encodes one frame*/
-int nb_encode(void *state, float *in, SpeexBits *bits);
+int nb_encode(void *state, short *in, SpeexBits *bits);
 
 
 /** Initializes decoder state*/
@@ -195,7 +195,7 @@ void *nb_decoder_init(SpeexMode *m);
 void nb_decoder_destroy(void *state);
 
 /** Decodes one frame*/
-int nb_decode(void *state, SpeexBits *bits, float *out);
+int nb_decode(void *state, SpeexBits *bits, short *out);
 
 /** ioctl-like function for controlling a narrowband encoder */
 int nb_encoder_ctl(void *state, int request, void *ptr);
