+++ libspeex/speex_stereo.h
@@ -51,10 +51,10 @@ typedef struct SpeexStereoState {
 #define SPEEX_STEREO_STATE_INIT {1,.5,1,1}
 
 /** Transforms a stereo frame into a mono frame and stores intensity stereo info in 'bits' */
-void speex_encode_stereo(float *data, int frame_size, SpeexBits *bits);
+void speex_encode_stereo(short *data, int frame_size, SpeexBits *bits);
 
 /** Transforms a mono frame into a stereo frame using intensity stereo info */
-void speex_decode_stereo(float *data, int frame_size, SpeexStereoState *stereo);
+void speex_decode_stereo(short *data, int frame_size, SpeexStereoState *stereo);
 
 /** Callback handler for intensity stereo info */
 int speex_std_stereo_request_handler(SpeexBits *bits, void *state, void *data);
