+++ libspeex/vbr.h
@@ -59,7 +59,7 @@ typedef struct VBRState {
 
 void vbr_init(VBRState *vbr);
 
-float vbr_analysis(VBRState *vbr, float *sig, int len, int pitch, float pitch_coef);
+float vbr_analysis(VBRState *vbr, short *sig, int len, int pitch, float pitch_coef);
 
 void vbr_destroy(VBRState *vbr);
